requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
autograd=1.6.2=pypi_0
autograd-gamma=0.5.0=pypi_0
autoprognosis=0.1.21=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
ca-certificates=2023.12.12=h06a4308_0
catboost=1.2.3=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=2.1.1=pypi_0
clarabel=0.7.1=pypi_0
cloudpickle=3.0.0=pypi_0
cmaes=0.10.0=pypi_0
colorlog=6.8.2=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.0=pypi_0
cvxopt=1.3.2=pypi_0
cvxpy=1.4.2=pypi_0
cycler=0.12.1=pypi_0
datasets=2.18.0=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.8=pypi_0
dm-tree=0.1.8=pypi_0
ecos=2.0.13=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fancyimpute=0.7.0=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.9.0=pypi_0
flatbuffers=24.3.7=pypi_0
fonttools=4.49.0=pypi_0
formulaic=1.0.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
future=1.0.0=pypi_0
gast=0.5.4=pypi_0
geomloss=0.2.6=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.4=pypi_0
hyperimpute=0.1.17=pypi_0
idna=3.4=pypi_0
importlib-metadata=7.0.2=pypi_0
importlib-resources=6.3.0=pypi_0
impyute=0.0.8=pypi_0
iniconfig=2.0.0=pypi_0
interface-meta=1.3.0=pypi_0
ipykernel=6.29.3=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.22=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.9.1=pypi_0
jupyter-lsp=2.2.4=pypi_0
jupyter-server=2.13.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.1.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.4=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
keras=3.0.5=pypi_0
kiwisolver=1.4.5=pypi_0
knnimpute=0.1.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lifelines=0.28.0=pypi_0
lightgbm=4.3.0=pypi_0
llvmlite=0.42.0=pypi_0
loguru=0.6.0=pypi_0
mako=1.3.2=pypi_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
miracle-imputation=0.1.5=pypi_0
mistune=3.0.2=pypi_0
ml-dtypes=0.3.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
namex=0.0.7=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.2=pypi_0
nbformat=5.10.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nose=1.3.7=pypi_0
notebook=7.1.1=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.59.0=pypi_0
numpy=1.26.4=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
optuna=3.1.0=pypi_0
osqp=0.6.5=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.2.0=pypi_0
plotly=5.19.0=pypi_0
pluggy=1.4.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=15.0.1=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11=2.11.1=pypi_0
pycparser=2.21=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pygments=2.17.2=pypi_0
pyod=1.1.3=pypi_0
pyparsing=3.1.2=pypi_0
pytest=8.1.1=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-graphviz=0.20.1=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
qdldl=0.1.7.post0=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
redis=5.0.3=pypi_0
referencing=0.33.0=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.11.4=pypi_0
scs=3.2.4.post1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.2.2=py39h06a4308_0
shap=0.45.0=pypi_0
six=1.16.0=pypi_0
slicer=0.0.7=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.28=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.3.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=2.1.0+cu121=pypi_0
torchaudio=2.1.0+cu121=pypi_0
torchvision=0.16.0+cu121=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
triton=2.1.0=pypi_0
types-python-dateutil=2.8.19.20240311=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.13=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.10=pypi_0
wrapt=1.16.0=pypi_0
xgboost=2.0.3=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.18.0=pypi_0
zlib=1.2.13=h5eee18b_0
